implicit numeric widening

+++ 2016/src/main/scala/Day9.scala
@@ -34,7 +34,7 @@ object Day9 extends SinglePuzzle[Int, Long] {
             nbRepeat.toInt * parseB(prefix) + parseB(suffix)
           case (None, None, s) =>
             val (prefix, suffix) = s.splitAt(s.indexOf('('))
-            if (prefix.isEmpty) suffix.length else prefix.length + parseB(suffix)
+            if (prefix.isEmpty) suffix.length.toLong else prefix.length + parseB(suffix)
         }
     }
   }
